fix: separate helper method from test file
